feat(app): add design system guidelines (#64)
#### What this PR does / why we need it:

Add design system guidelines.

**KEYCHAIN DESIGN SYSTEM**
A collection of open source rules, principles, and constraints that
govern how we design and build Levain Keychain. It is a living document
that will grow and evolve as we add more features and functionality to
Keychain. As a design system, it provides a common language and
vocabulary to build cohesive, accessible, responsive, and efficient
experiences.