chore(app): introduce PDD and `./app/components` directory (#96)
#### What this PR does / why we need it:

PDD is a paradigm that we use to structure our codebase. It's a paradigm
where code is organized by context, not by technology or false
separation of concerns. It's a paradigm that enables us to scale our
codebase without sacrificing the ability to adapt to iterations, to
refactor, and to maintain.

This PR also updated a few design styles.